[2.0.0-SNAPSHOT] Package refactoring README.md updated
