+++ src/main/java/io/goodforgod/dummymapper/AvroSchemaJacksonAction.java
@@ -1,7 +1,7 @@
 package io.goodforgod.dummymapper;
 
-import io.goodforgod.dummymapper.mapper.MarkerMapper;
-import io.goodforgod.dummymapper.mapper.impl.AvroJacksonMapper;
+import io.goodforgod.dummymapper.marker.MarkerMapper;
+import io.goodforgod.dummymapper.marker.mapper.AvroJacksonMapper;
 import io.goodforgod.dummymapper.ui.config.AvroJacksonConfig;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
