+++ src/main/java/io/goodforgod/dummymapper/JsonSchemaAction.java
@@ -1,6 +1,6 @@
 package io.goodforgod.dummymapper;
 
-import io.goodforgod.dummymapper.mapper.impl.JsonSchemaMapper;
+import io.goodforgod.dummymapper.marker.mapper.JsonSchemaMapper;
 import io.goodforgod.dummymapper.ui.config.JsonSchemaConfig;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
