+++ src/main/java/io/goodforgod/dummymapper/MapperAction.java
@@ -13,9 +13,9 @@ import io.dummymaker.util.StringUtils;
 import io.goodforgod.dummymapper.error.MapperException;
 import io.goodforgod.dummymapper.error.PsiKindException;
 import io.goodforgod.dummymapper.error.UnsupportedPsiFileException;
-import io.goodforgod.dummymapper.mapper.MarkerMapper;
+import io.goodforgod.dummymapper.marker.MarkerMapper;
 import io.goodforgod.dummymapper.marker.RawMarker;
-import io.goodforgod.dummymapper.scanner.PsiClassScanner;
+import io.goodforgod.dummymapper.service.PsiClassScanner;
 import io.goodforgod.dummymapper.ui.ConfigDialog;
 import io.goodforgod.dummymapper.ui.config.IConfig;
 import io.goodforgod.dummymapper.util.IdeaUtils;
