+++ src/main/java/io/goodforgod/dummymapper/marker/filter/ExcludeSetterAnnotationFilter.java
@@ -1,6 +1,6 @@
-package io.goodforgod.dummymapper.filter.impl;
+package io.goodforgod.dummymapper.marker.filter;
 
-import io.goodforgod.dummymapper.model.AnnotationMarker;
+import io.goodforgod.dummymapper.marker.AnnotationMarker;
 import java.util.function.Predicate;
 
 /**
