+++ src/main/java/io/goodforgod/dummymapper/marker/filter/GraphQLQueryFilter.java
@@ -1,7 +1,7 @@
-package io.goodforgod.dummymapper.filter.impl;
+package io.goodforgod.dummymapper.marker.filter;
 
+import io.goodforgod.dummymapper.marker.AnnotationMarker;
 import io.goodforgod.dummymapper.marker.RawMarker;
-import io.goodforgod.dummymapper.model.AnnotationMarkerBuilder;
 import io.leangen.graphql.annotations.GraphQLNonNull;
 import io.leangen.graphql.annotations.GraphQLQuery;
 import org.jetbrains.annotations.NotNull;
@@ -21,7 +21,7 @@ public class GraphQLQueryFilter extends BaseFilter {
         marker.getStructure().entrySet().stream()
                 .filter(e -> e.getValue().getAnnotations().stream()
                         .noneMatch(a -> a.named(GraphQLQuery.class)))
-                .forEach(e -> e.getValue().addAnnotation(AnnotationMarkerBuilder.get()
+                .forEach(e -> e.getValue().addAnnotation(AnnotationMarker.builder()
                         .ofField()
                         .withName(GraphQLQuery.class)
                         .build()));
