+++ src/main/java/io/goodforgod/dummymapper/marker/mapper/AvroJacksonMapper.java
@@ -1,4 +1,4 @@
-package io.goodforgod.dummymapper.mapper.impl;
+package io.goodforgod.dummymapper.marker.mapper;
 
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -8,14 +8,14 @@ import com.fasterxml.jackson.dataformat.avro.schema.AvroSchemaGenerator;
 import com.intellij.psi.PsiJavaFile;
 import io.goodforgod.dummymapper.error.ExternalException;
 import io.goodforgod.dummymapper.error.MapperException;
-import io.goodforgod.dummymapper.filter.MarkerFilter;
-import io.goodforgod.dummymapper.filter.impl.AvroFilter;
-import io.goodforgod.dummymapper.filter.impl.EmptyMarkerFilter;
-import io.goodforgod.dummymapper.filter.impl.ExcludeSetterAnnotationFilter;
-import io.goodforgod.dummymapper.filter.impl.JacksonPropertyFilter;
-import io.goodforgod.dummymapper.mapper.MarkerMapper;
+import io.goodforgod.dummymapper.marker.MarkerFilter;
+import io.goodforgod.dummymapper.marker.MarkerMapper;
 import io.goodforgod.dummymapper.marker.RawMarker;
-import io.goodforgod.dummymapper.service.ClassFactory;
+import io.goodforgod.dummymapper.marker.filter.AvroFilter;
+import io.goodforgod.dummymapper.marker.filter.EmptyMarkerFilter;
+import io.goodforgod.dummymapper.marker.filter.ExcludeSetterAnnotationFilter;
+import io.goodforgod.dummymapper.marker.filter.JacksonPropertyFilter;
+import io.goodforgod.dummymapper.service.AssistClassFactory;
 import io.goodforgod.dummymapper.ui.config.AvroJacksonConfig;
 import java.util.Optional;
 import org.apache.avro.Schema;
@@ -36,7 +36,7 @@ public class AvroJacksonMapper implements MarkerMapper<AvroJacksonConfig> {
     private final MarkerFilter avroFilter = new AvroFilter();
     private final MarkerFilter requiredFieldFilter = new JacksonPropertyFilter();
     private final MarkerFilter annotationFilter = new ExcludeSetterAnnotationFilter();
-    private final ObjectMapper avroMapper = AbstractJsonJacksonMapper.configure(new ObjectMapper(new AvroFactory()));
+    private final ObjectMapper avroMapper = ObjectMapperUtils.configure(new ObjectMapper(new AvroFactory()));
 
     @NotNull
     @Override
@@ -54,7 +54,7 @@ public class AvroJacksonMapper implements MarkerMapper<AvroJacksonConfig> {
             if (filtered.isEmpty())
                 return "";
 
-            final Class<?> target = ClassFactory.build(filtered);
+            final Class<?> target = AssistClassFactory.build(filtered);
 
             final AvroSchemaGenerator generator = new AvroSchemaGenerator();
             avroMapper.acceptJsonFormatVisitor(target, generator);
