+++ src/main/java/io/goodforgod/dummymapper/marker/mapper/JsonSchemaMapper.java
@@ -1,14 +1,14 @@
-package io.goodforgod.dummymapper.mapper.impl;
+package io.goodforgod.dummymapper.marker.mapper;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.github.victools.jsonschema.generator.*;
 import com.intellij.psi.PsiJavaFile;
 import io.goodforgod.dummymapper.error.ExternalException;
-import io.goodforgod.dummymapper.filter.MarkerFilter;
-import io.goodforgod.dummymapper.filter.impl.EmptyMarkerFilter;
-import io.goodforgod.dummymapper.mapper.MarkerMapper;
+import io.goodforgod.dummymapper.marker.MarkerFilter;
+import io.goodforgod.dummymapper.marker.MarkerMapper;
 import io.goodforgod.dummymapper.marker.RawMarker;
-import io.goodforgod.dummymapper.service.ClassFactory;
+import io.goodforgod.dummymapper.marker.filter.EmptyMarkerFilter;
+import io.goodforgod.dummymapper.service.AssistClassFactory;
 import io.goodforgod.dummymapper.ui.config.JsonSchemaConfig;
 import java.util.Optional;
 import org.jetbrains.annotations.NotNull;
@@ -35,7 +35,7 @@ public class JsonSchemaMapper implements MarkerMapper<JsonSchemaConfig> {
             if (filtered.isEmpty())
                 return "";
 
-            final Class<?> target = ClassFactory.build(filtered);
+            final Class<?> target = AssistClassFactory.build(filtered);
 
             final SchemaVersion version = config.getSchemaVersion();
             final SchemaGeneratorConfig generatorConfig = new SchemaGeneratorConfigBuilder(version, OptionPreset.PLAIN_JSON)
